+++ src/main/java/cz/cvut/fit/tjv/popovle1/semestral/crud_gamedev/converter/GameConverter.java
@@ -14,7 +14,7 @@ public class GameConverter {
 //    }
 
     public static GameDTO toDTO(Game game) {
-        List<Long> studiosIds = new ArrayList<>();
+        List<Integer> studiosIds = new ArrayList<>();
         if (game.getStudios() != null)
             for (Studio studio : game.getStudios())
                 studiosIds.add(studio.getId());
