+++ src/main/java/cz/cvut/fit/tjv/popovle1/semestral/crud_gamedev/dto/GameDTO.java
@@ -3,19 +3,19 @@ package cz.cvut.fit.tjv.popovle1.semestral.crud_gamedev.dto;
 import java.util.List;
 
 public class GameDTO {
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String genre;
-    private final List<Long> studiosIds;
+    private final List<Integer> studiosIds;
 
-    public GameDTO(Long id, String name, String genre, List<Long> studiosIds) {
+    public GameDTO(Integer id, String name, String genre, List<Integer> studiosIds) {
         this.id = id;
         this.name = name;
         this.genre = genre;
         this.studiosIds = studiosIds;
     }
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
 
     public String getName() {
         return name;
@@ -25,7 +25,7 @@ public class GameDTO {
         return genre;
     }
 
-    public List<Long> getStudiosIds() {
+    public List<Integer> getStudiosIds() {
         return studiosIds;
     }
 }
